+++ cmd/flags/proposer.go
@@ -55,6 +55,10 @@ var (
 		Value:    1,
 		Category: proposerCategory,
 	}
+	ProposeBlockTxGasLimit = &cli.Uint64Flag{
+		Name:     "proposeBlockTxGasLimit",
+		Category: proposerCategory,
+	}
 )
 
 // All proposer flags.
@@ -68,4 +72,5 @@ var ProposerFlags = MergeFlags(CommonFlags, []cli.Flag{
 	ProposeEmptyBlocksInterval,
 	MinBlockGasLimit,
 	MaxProposedTxListsPerEpoch,
+	ProposeBlockTxGasLimit,
 })