+++ metrics/metrics.go
@@ -42,6 +42,8 @@ var (
 	ProverSentInvalidProofCounter     = metrics.NewRegisteredCounter("prover/proof/invalid/sent", nil)
 	ProverReceivedProposedBlockGauge  = metrics.NewRegisteredGauge("prover/proposed/received", nil)
 	ProverReceivedProvenBlockGauge    = metrics.NewRegisteredGauge("prover/proven/received", nil)
+	ProverAllProofRewardGauge         = metrics.NewRegisteredGauge("prover/allProofReward", nil)
+	ProverNormalProofRewardGauge      = metrics.NewRegisteredGauge("prover/normalProofReward", nil)
 )
 
 // Serve starts the metrics server on the given address, will be closed when the given