+++ pkg/rpc/methods.go
@@ -25,8 +25,10 @@ var (
 	errSyncing = errors.New("syncing")
 	// syncProgressRecheckDelay is the time delay of rechecking the L2 execution engine's sync progress again,
 	// if the previous check failed.
-	syncProgressRecheckDelay = 12 * time.Second
-	minTxGasLimit            = 21000
+	syncProgressRecheckDelay    = 12 * time.Second
+	waitL1OriginPollingInterval = 3 * time.Second
+	defaultWaitL1OriginTimeout  = 3 * time.Minute
+	minTxGasLimit               = 21000
 )
 
 // ensureGenesisMatched fetches the L2 genesis block from TaikoL1 contract,
@@ -74,8 +76,8 @@ func (c *Client) ensureGenesisMatched(ctx context.Context) error {
 	return nil
 }
 
-// WaitTillL2Synced keeps waiting until the L2 execution engine is fully synced.
-func (c *Client) WaitTillL2Synced(ctx context.Context) error {
+// WaitTillL2ExecutionEngineSynced keeps waiting until the L2 execution engine is fully synced.
+func (c *Client) WaitTillL2ExecutionEngineSynced(ctx context.Context) error {
 	return backoff.Retry(
 		func() error {
 			if ctx.Err() != nil {
@@ -172,29 +174,38 @@ func (c *Client) WaitL1Origin(ctx context.Context, blockID *big.Int) (*rawdb.L1O
 		err      error
 	)
 
-	ticker := time.NewTicker(time.Second)
+	ticker := time.NewTicker(waitL1OriginPollingInterval)
 	defer ticker.Stop()
 
-	log.Debug("Start fetching L1Origin from L2 execution engine", "blockID", blockID)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		case <-ticker.C:
-			l1Origin, err = c.L2.L1OriginByID(ctx, blockID)
-			if err != nil {
-				log.Warn("Failed to fetch L1Origin from L2 execution engine", "blockID", blockID, "error", err)
-				continue
-			}
-
-			if l1Origin == nil {
-				continue
-			}
-
-			return l1Origin, nil
-		}
+	var (
+		ctxWithTimeout = ctx
+		cancel         context.CancelFunc
+	)
+	if _, ok := ctx.Deadline(); !ok {
+		ctxWithTimeout, cancel = context.WithTimeout(ctx, defaultWaitL1OriginTimeout)
+		defer cancel()
 	}
+
+	log.Debug("Start fetching L1Origin from L2 execution engine", "blockID", blockID)
+	for ; true; <-ticker.C {
+		if ctxWithTimeout.Err() != nil {
+			return nil, ctx.Err()
+		}
+
+		l1Origin, err = c.L2.L1OriginByID(ctxWithTimeout, blockID)
+		if err != nil {
+			log.Warn("L1Origin from L2 execution engine not found, keep retrying", "blockID", blockID, "error", err)
+			continue
+		}
+
+		if l1Origin == nil {
+			continue
+		}
+
+		return l1Origin, nil
+	}
+
+	return nil, fmt.Errorf("failed to fetch L1Origin from L2 execution engine, blockID: %d", blockID)
 }
 
 // GetPoolContent fetches the transactions list from L2 execution engine's transactions pool with given
@@ -345,12 +356,18 @@ func (c *Client) CheckL1Reorg(ctx context.Context, blockID *big.Int) (bool, *typ
 
 		l1Origin, err := c.L2.L1OriginByID(ctx, blockID)
 		if err != nil {
+			// If the L2 EE is just synced through P2P, there is a chance that the EE do not have
+			// the chain head L1Origin information recorded.
+			if err.Error() == ethereum.NotFound.Error() {
+				log.Info("L1Origin not found", "blockID", blockID)
+				return false, nil, nil, nil
+			}
 			return false, nil, nil, err
 		}
 
 		l1Header, err := c.L1.HeaderByNumber(ctx, l1Origin.L1BlockHeight)
 		if err != nil {
-			if errors.Is(err, ethereum.NotFound) {
+			if err.Error() == ethereum.NotFound.Error() {
 				continue
 			}
 			return false, nil, nil, fmt.Errorf("failed to fetch L1 header (%d): %w", l1Origin.L1BlockHeight, err)