+++ pkg/rpc/utils_test.go
@@ -6,21 +6,9 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/require"
 )
-
-func TestWaitConfirmations(t *testing.T) {
-	client := newTestClient(t)
-
-	l1Head, err := client.L1.BlockNumber(context.Background())
-	require.Nil(t, err)
-
-	require.Nil(t, client.L1RawRPC.CallContext(context.Background(), nil, "hardhat_mine", hexutil.EncodeUint64(4)))
-
-	require.Nil(t, WaitConfirmations(context.Background(), client.L1, 4, l1Head))
-}
 
 func TestWaitReceiptTimeout(t *testing.T) {
 	client := newTestClient(t)
@@ -33,17 +21,6 @@ func TestWaitReceiptTimeout(t *testing.T) {
 	)
 
 	require.ErrorContains(t, err, "context deadline exceeded")
-}
-
-func TestGetReceiptsByBlock(t *testing.T) {
-	client := newTestClient(t)
-
-	l1Genesis, err := client.L1.BlockByNumber(context.Background(), common.Big0)
-	require.Nil(t, err)
-
-	receipts, err := GetReceiptsByBlock(context.Background(), client.L1RawRPC, l1Genesis)
-	require.Nil(t, err)
-	require.Empty(t, receipts)
 }
 
 func TestSetHead(t *testing.T) {