+++ proposer/proposer.go
@@ -27,7 +27,8 @@ import (
 )
 
 var (
-	errNoNewTxs = errors.New("no new transactions")
+	errNoNewTxs        = errors.New("no new transactions")
+	waitReceiptTimeout = 1 * time.Minute
 )
 
 // Proposer keep proposing new transactions from L2 execution engine's tx pool at a fixed interval.
@@ -48,6 +49,7 @@ type Proposer struct {
 	locals                     []common.Address
 	minBlockGasLimit           *uint64
 	maxProposedTxListsPerEpoch uint64
+	proposeBlockTxGasLimit     *uint64
 
 	// Protocol configurations
 	protocolConfigs *bindings.TaikoDataConfig
@@ -77,6 +79,7 @@ func InitFromConfig(ctx context.Context, p *Proposer, cfg *Config) (err error) {
 	p.l2SuggestedFeeRecipient = cfg.L2SuggestedFeeRecipient
 	p.proposingInterval = cfg.ProposeInterval
 	p.proposeEmptyBlocksInterval = cfg.ProposeEmptyBlocksInterval
+	p.proposeBlockTxGasLimit = cfg.ProposeBlockTxGasLimit
 	p.wg = sync.WaitGroup{}
 	p.locals = cfg.LocalAddresses
 	p.commitSlot = cfg.CommitSlot
@@ -180,7 +183,7 @@ func (p *Proposer) ProposeOp(ctx context.Context) error {
 	}
 
 	// Wait until L2 execution engine is synced at first.
-	if err := p.rpc.WaitTillL2Synced(ctx); err != nil {
+	if err := p.rpc.WaitTillL2ExecutionEngineSynced(ctx); err != nil {
 		return fmt.Errorf("failed to wait until L2 execution engine synced: %w", err)
 	}
 
@@ -295,13 +298,19 @@ func (p *Proposer) ProposeTxList(
 	if nonce != nil {
 		opts.Nonce = new(big.Int).SetUint64(*nonce)
 	}
+	if p.proposeBlockTxGasLimit != nil {
+		opts.GasLimit = *p.proposeBlockTxGasLimit
+	}
 
 	proposeTx, err := p.rpc.TaikoL1.ProposeBlock(opts, inputs, txListBytes)
 	if err != nil {
 		return encoding.TryParsingCustomError(err)
 	}
 
-	if _, err := rpc.WaitReceipt(ctx, p.rpc.L1, proposeTx); err != nil {
+	ctxWithTimeout, cancel := context.WithTimeout(ctx, waitReceiptTimeout)
+	defer cancel()
+
+	if _, err := rpc.WaitReceipt(ctxWithTimeout, p.rpc.L1, proposeTx); err != nil {
 		return err
 	}
 
@@ -335,8 +344,8 @@ func (p *Proposer) updateProposingTicker() {
 	if p.proposingInterval != nil {
 		duration = *p.proposingInterval
 	} else {
-		// Random number between 12 - 60
-		randomSeconds := rand.Intn(60-11) + 12
+		// Random number between 12 - 120
+		randomSeconds := rand.Intn(120-11) + 12
 		duration = time.Duration(randomSeconds) * time.Second
 	}