+++ prover/proof_submitter/valid_proof_submitter.go
@@ -119,12 +119,12 @@ func (s *ValidProofSubmitter) RequestProof(ctx context.Context, event *bindings.
 	// Get the header of the block to prove from L2 execution engine.
 	block, err := s.rpc.L2.BlockByHash(ctx, l1Origin.L2BlockHash)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get the current L2 block by hash (%s): %w", l1Origin.L2BlockHash, err)
 	}
 
 	parent, err := s.rpc.L2.BlockByHash(ctx, block.ParentHash())
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get the L2 parent block by hash (%s): %w", block.ParentHash(), err)
 	}
 
 	blockInfo, err := s.rpc.TaikoL1.GetBlock(nil, event.Id)
@@ -159,7 +159,7 @@ func (s *ValidProofSubmitter) RequestProof(ctx context.Context, event *bindings.
 	}
 
 	if err := s.proofProducer.RequestProof(ctx, opts, event.Id, &event.Meta, block.Header(), s.resultCh); err != nil {
-		return err
+		return fmt.Errorf("failed to request proof (id: %d): %w", event.Id, err)
 	}
 
 	metrics.ProverQueuedProofCounter.Inc(1)
@@ -286,13 +286,6 @@ func (s *ValidProofSubmitter) SubmitProof(
 
 		return err
 	}
-
-	log.Info(
-		"💰 Your block proof was accepted",
-		"blockID", proofWithHeader.BlockID,
-		"hash", block.Hash(), "height", block.Number(),
-		"transactions", block.Transactions().Len(),
-	)
 
 	metrics.ProverSentProofCounter.Inc(1)
 	metrics.ProverSentValidProofCounter.Inc(1)